+++ assistant_gazebo/maps/small_map.yaml
@@ -1,0 +1,7 @@
+image: /home/user/workspace/turtlebot3/base/src/assistant/assistant_gazebo/maps/small_map.pgm
+resolution: 0.050000
+origin: [-51.224998, -51.224998, 0.000000]
+negate: 0
+occupied_thresh: 0.65
+free_thresh: 0.196
+
