+++ assistant_navigation/maps/ucu_test.yaml
@@ -1,0 +1,6 @@
+image: ucu_test.pgm
+resolution: 0.050000
+origin: [-100.000000, -100.000000, 0.000000]
+negate: 0
+occupied_thresh: 0.65
+free_thresh: 0.196
